+++ src/ladb_toolbox/js/plugins/jquery.ladb.tab-about.js
@@ -8,35 +8,11 @@
         this.options = options;
         this.$element = $(element);
         this.toolbox = toolbox;
-
-        this.$btnCheckForUpdate = $('#ladb_btn_check_for_update', this.$header);
-        this.$btnUpgrade = $('#ladb_btn_upgrade', this.$header);
-
     };
 
     LadbTabAbout.DEFAULTS = {};
 
-    LadbTabAbout.prototype.bind = function () {
-        var that = this;
-
-        // Bind buttons
-        this.$btnCheckForUpdate.on('click', function () {
-
-            rubyCallCommand('check_for_update', {}, function(response) {});
-
-            this.blur();
-        });
-        this.$btnUpgrade.on('click', function () {
-
-            rubyCallCommand('upgrade', {}, function(response) {});
-
-            this.blur();
-        });
-
-    };
-
     LadbTabAbout.prototype.init = function () {
-        this.bind();
     };
 
 
@@ -50,7 +26,7 @@
             var options = $.extend({}, LadbTabAbout.DEFAULTS, $this.data(), typeof option == 'object' && option);
 
             if (!data) {
-                if (options.toolbox == undefined) {
+                if (undefined === options.toolbox) {
                     throw 'toolbox option is mandatory.';
                 }
                 $this.data('ladb.tabAbout', (data = new LadbTabAbout(this, options, options.toolbox)));